Reuse generic hard sphere implementation in PeTS model
