+++ src/pets/eos/mod.rs
@@ -1,4 +1,5 @@
 use super::parameters::PetsParameters;
+use crate::hard_sphere::HardSphere;
 use feos_core::joback::Joback;
 use feos_core::parameter::Parameter;
 use feos_core::{
@@ -11,10 +12,8 @@ use std::f64::consts::{FRAC_PI_6, PI};
 use std::rc::Rc;
 
 pub(crate) mod dispersion;
-pub(crate) mod hard_sphere;
 mod qspr;
 use dispersion::Dispersion;
-use hard_sphere::HardSphere;
 use qspr::QSPR;
 
 #[allow(clippy::upper_case_acronyms)]
@@ -24,8 +23,8 @@ enum IdealGasContributions {
 }
 
 /// Configuration options for the PeTS equation of state and Helmholtz energy functional.
-/// 
-/// The maximum packing fraction is used to infer initial values 
+///
+/// The maximum packing fraction is used to infer initial values
 /// for routines that depend on starting values for the system density.
 #[derive(Copy, Clone)]
 pub struct PetsOptions {
@@ -56,9 +55,7 @@ impl Pets {
     /// PeTS equation of state with provided options.
     pub fn with_options(parameters: Rc<PetsParameters>, options: PetsOptions) -> Self {
         let contributions: Vec<Box<dyn HelmholtzEnergy>> = vec![
-            Box::new(HardSphere {
-                parameters: parameters.clone(),
-            }),
+            Box::new(HardSphere::new(&parameters)),
             Box::new(Dispersion {
                 parameters: parameters.clone(),
             }),
@@ -332,7 +329,9 @@ mod tests {
         argon_krypton_parameters, argon_parameters, krypton_parameters,
     };
     use approx::assert_relative_eq;
-    use feos_core::{Contributions, DensityInitialization, PhaseEquilibrium, State};
+    use feos_core::{
+        Contributions, DensityInitialization, HelmholtzEnergyDual, PhaseEquilibrium, State, StateHD,
+    };
     use ndarray::arr1;
     use quantity::si::{BAR, KELVIN, METER, PASCAL, RGAS};
 
@@ -366,6 +365,25 @@ mod tests {
             s.pressure(Contributions::Total),
             epsilon = 1e-10
         );
+    }
+
+    #[test]
+    fn hard_sphere_mix() {
+        let c1 = HardSphere::new(&argon_parameters());
+        let c2 = HardSphere::new(&krypton_parameters());
+        let c12 = HardSphere::new(&argon_krypton_parameters());
+        let t = 250.0;
+        let v = 2.5e28;
+        let n = 1.0;
+        let s = StateHD::new(t, v, arr1(&[n]));
+        let a1 = c1.helmholtz_energy(&s);
+        let a2 = c2.helmholtz_energy(&s);
+        let s1m = StateHD::new(t, v, arr1(&[n, 0.0]));
+        let a1m = c12.helmholtz_energy(&s1m);
+        let s2m = StateHD::new(t, v, arr1(&[0.0, n]));
+        let a2m = c12.helmholtz_energy(&s2m);
+        assert_relative_eq!(a1, a1m, epsilon = 1e-14);
+        assert_relative_eq!(a2, a2m, epsilon = 1e-14);
     }
 
     #[test]
